cloud function heuristic/requirements.txt
# Function dependencies, for example:
# package>=version
flask-cors
